Move ExampleCache* to cache_test.go
